requirements.txt
aiofiles
aiohttp
aiohttp_session
cryptography
pyrofork
python-dotenv
parse-torrent-title
tgcrypto
uvloop
pymongo
requests